+++ src/common/processing_linux.c
@@ -165,6 +165,18 @@ void ffProcessGetInfoLinux(pid_t pid, FFstrbuf* processName, FFstrbuf* exe, cons
 
             do arg0++; while (*arg0 == '\0');
             assert(arg0 < procArgs2 + len);
+
+            {
+                // #977
+                char* p = (char*) strrchr(arg0, '/');
+                if (p)
+                {
+                    p++;
+                    if (ffStrStartsWithIgnCase(p, "python")) // /opt/homebrew/Cellar/python@3.12/3.12.3/Frameworks/Python.framework/Versions/3.12/Resources/Python.app/Contents/MacOS/Python /Users/carter/.local/bin/xonsh
+                        arg0 = p + strlen(p) + 1;
+                }
+            }
+
             if (*arg0 == '-') arg0++; // Login shells
 
             ffStrbufSetS(exe, arg0);