+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Features:
     * See `fastfetch -h format` for detail
 * Add option `--color-separator` to set the color of key-value separators
 * Support Guix package manager count (#792, Packages, Linux)
+* Improve python based shell detection (#977, Shell, macOS)
 
 Bugfixes:
 * Don't detect `.conf` files in `--list-config-paths`